Move parser logic to components
Changed import path in uploadFile.js
Updated steps documentation in README.md
