+++ monolithstarter-static/src/components/parser.js
